#  Async Runtime Performance Optimization: Fix CPU Thrashing Issues